Add basic  PoolAllocator alongside gtests for it
This is primitive implementation of custom memory management tools.
PoolAllocator takes responsibility for allocating memory for the entire program (master allocation) as well as for smaller chunks (that are allocated from master allocation) or for any objects. This allocator also tracks all references to allocated blocks and nullifies them if a block is released.
TMemoryHandle is also introduced as mixed version of shared and weak smart pointers.